Arm backend: Move quant util functions to closer to usage
The following functions are only used in quantization_annotator and can
therefore be moved from arm_quantizer_utils.py to
quantization_annotator.py:

    * is_large_scalar
    * is_non_float_tensor
    * get_node_target

Additionally, is_ok_for_quantization is removed. It combined the
is_large_scalar and is_non_float_tensor checks into one, which is now
done directly where is_ok_for_quantization was used.
